feat(BuySell): move medium component and rendering logic out of template
