+++ packages/blockchain-wallet-v4-frontend/src/data/modules/coinify/actionTypes.js
@@ -6,3 +6,5 @@ export const COINIFY_NEXT_STEP = '@COMPONENT.COINIFY_NEXT_STEP'
 export const COINIFY_CLEAR_SIGNUP_ERROR = '@COMPONENT.COINIFY_CLEAR_SIGNUP_ERROR'
 
 export const COINIFY_SAVE_QUOTE = '@COMPONENT.COINIFY_SAVE_QUOTE'
+
+export const COINIFY_SAVE_MEDIUM = '@COMPONENT.COINIFY_SAVE_MEDIUM'
