+++ packages/blockchain-wallet-v4-frontend/src/data/modules/coinify/reducers.js
@@ -19,6 +19,9 @@ const coinify = (state = INITIAL_STATE, action) => {
     case AT.COINIFY_SAVE_QUOTE: {
       return assoc('quote', payload, state)
     }
+    case AT.COINIFY_SAVE_MEDIUM: {
+      return assoc('medium', payload, state)
+    }
     default:
       return state
   }
